UploadFileVec uses FileVec instead of ByteVec; most code gets removed.
FileVec can have it's _len set during close.
Remove getChunkSz back to CHUNK_SZ so the compiler gets a constant.
Remove espc() back to _espc so no v-call in inner loops.
Use Arrays.copyOf instead byte-by-byte copy.
Use FileVec.CHUNK_SZ instead of Vec.CHUNK_SZ.
Shrink CHUNK_SZ to an int.
Restore close() behavior.

+++ src/main/java/water/fvec/NewChunk.java
@@ -120,7 +120,7 @@ public class NewChunk extends Chunk {
       int nzcnt=0;
       for( int i=0; i<_len; i++ ) {
         if( _ls[i]!=0 ) nzcnt++;
-        if( _xs[i]!=0 ) { nzcnt = (int)Vec.CHUNK_SZ; break; } // Only non-specials sparse
+        if( _xs[i]!=0 ) { nzcnt = Vec.CHUNK_SZ; break; } // Only non-specials sparse
       }
       if( _len >= 32 && nzcnt*8 <= _len ) { // Heuristic for sparseness
         _len=0;
